Add invite subscriptions
I can't figure out where to configure interceptors to view headers for the
initial request that triggers a subscription.

Currently passing it in as an arg to the subscription, but that's pretty hacky.

Connection params are documented, but those are harder to set in graphiql, and
then there are two auth methods instead of one.